Removed configurations from VC
