+++ autolayout/Core/Components/Views/DollarView.swift
@@ -8,7 +8,7 @@
 
 import UIKit
 
-class DollarView: UIView {
+final class DollarView: UIView {
 
     override init(frame: CGRect) {
         super.init(frame: frame)
