+++ autolayout/Core/Components/Views/ScrollView.swift
@@ -10,10 +10,19 @@ import UIKit
 
 final class ScrollView: UIScrollView {
 
+    // MARK: - Constants
+
+    let containerView = ContainerView()
+
+    // MARK: - Initializers
+
     override init(frame: CGRect) {
         super.init(frame: frame)
 
         translatesAutoresizingMaskIntoConstraints = false
+
+        setupView()
+        setupContainerView()
     }
 
     @available(*, unavailable)
@@ -35,3 +44,23 @@ extension ScrollView {
     }
 }
 
+extension ScrollView {
+
+    func setupView() {
+        self.addSubview(containerView)
+    }
+
+    func setupContainerView() {
+        let constraint = containerView.heightAnchor.constraint(equalTo: self.heightAnchor)
+        constraint.priority = .defaultLow
+
+        NSLayoutConstraint.activate([
+            containerView.leadingAnchor.constraint(equalTo: self.leadingAnchor),
+            containerView.trailingAnchor.constraint(equalTo: self.trailingAnchor),
+            containerView.topAnchor.constraint(equalTo: self.topAnchor),
+            containerView.bottomAnchor.constraint(equalTo: self.bottomAnchor),
+            containerView.widthAnchor.constraint(equalTo: self.widthAnchor),
+            constraint
+        ])
+    }
+}
