+++ autolayout/Core/Components/Views/WidgetViews/CellularWidgetView.swift
@@ -8,19 +8,11 @@
 
 import UIKit
 
-final class CellularWidgetView: UIView {
+final class CellularWidgetView: WidgetView {
 
     override init(frame: CGRect) {
         super.init(frame: frame)
-
-        translatesAutoresizingMaskIntoConstraints = false
     }
-
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
 }
 
 // MARK: - Configuration
@@ -29,11 +21,9 @@ extension CellularWidgetView {
 
     struct Model {
         let backgroundColor: UIColor
-        let cornerRadius: CGFloat
     }
 
     func configure(with model: Model) {
         backgroundColor = model.backgroundColor
-        layer.cornerRadius = model.cornerRadius
     }
 }
