+++ autolayout/Core/Components/Views/WidgetViews/SchedulerWidgetView.swift
@@ -1,0 +1,17 @@
+//
+//  SchedulerWidgetView.swift
+//  autolayout
+//
+//  Created by AndUser on 02.12.2020.
+//  Copyright © 2020 AndUser. All rights reserved.
+//
+
+import UIKit
+
+final class SchedulerWidgetView: WidgetView {
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+    }
+
+}
